fix: No server-side fetch request to /lti
Fixes https://github.com/serlo/ece-as-a-service/issues/156
