+++ .env-docker
@@ -1,9 +1,3 @@
 MONGODB_URL=mongodb://host.docker.internal:27017/admin
 PLATFORM_JWK_SET=http://host.docker.internal:8100/edu-sharing/rest/lti/v13/jwks
 EDUSHARING_NETWORK_HOST=host.docker.internal:8100
-
-# In the local docker run we need to use `ltijs` in dev mode since we run the
-# server via HTTP and not HTTPS. However we run the local docker build with
-# NODE_ENV=production. Thus we need another way to detect whether we are in a
-# local docker run.
-IS_LOCAL_DOCKER_RUN=true
